Added steam API scraper.
It runs and works for public profiles. Input is steam login id. For
Andrew and his 740+ plus games, it runs for like 10 minutes before it
returns. For me and my like 6 games its pretty quick. Feel free to tear
it apart and change anything. Sorry Steven if it is bad python.